requirements.txt
fastapi
uvicorn
jinja2
requests
difflib2
psycopg2-binary
